Address comments by @altendky
